examples: fix README & relocating files and directories
