+++ examples/flow_field_prediction/ppdiffusion/README.md
