+++ examples/flow_field_prediction/ppdiffusion/datamodules/physical.py
