--- examples/ppdiffusion/metrics/__init__.py
@@ -1,8 +1,0 @@
-from __future__ import annotations
-
-
-from .ensemble import EnsembleMetrics  # isort:skip
-
-__all__ = [
-    "EnsembleMetrics",
-]
