+++ ppcfd/models/ppdiffusion/metrics/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+
+from ppcfd.models.ppdiffusion.metrics.ensemble import EnsembleMetrics  # isort:skip
+
+__all__ = [
+    "EnsembleMetrics",
+]
